requirements.txt
Django==5.0.1
transformers==4.37.2
requests==2.31.0
Pillow==10.2.0
torch==2.1.2
accelerate==0.26.1
bitsandbytes==0.42.0